Add a Dockerfile
This allows us to deploy River5 as a Docker container, an isolated,
controlled sandbox. The container image can then be deployed to a
variety of services including Amazon's EC2 Container Service and
Docker Cloud.

To build a container for your own river, you'll need to install Docker.
Once you've got that working:

1. Open a terminal
2. Navigate to your river5 folder
3. Run `docker build -t river5:latest .`

This creates an image, which you then run to create a container
(essentially the Node.js process in a sandbox):

```
docker run -p 1337:1337 river5:latest
```

A few notes:

* This Dockerfile assumes that your `config.js` is set to use port 1337, which is the default.
* You will need to configure your river before you build the container. The `docker build` process then encapsulates your config inside the image.
* Getting Docker working is not always trivial, though there are reasonable installers for common OSs at docker.com.